% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sent.R
\name{db_sent}
\alias{db_sent}
\title{Sliping sentences from txt object.}
\usage{
db_sent(txt = txt, minL = 50, maxL = 800)
}
\arguments{
\item{txt}{A txt object containg characters.}

\item{minL}{Filter the sentence by length longer than minL.}

\item{maxL}{Filter the sentence by length shorter than maxL.}
}
\description{
Sliping sentences from txt object.
}
\examples{
db_sci <- list(dest = "PATH/Mendeley_PDF/", myfiles = NULL, sents = NULL) # build a empty list comprising 3 elements. Assign path of the files saved by mendeley
db_sci$myfiles <- db_sci$dest \%>\% all_pdf # save all the pdf file names in this element
pdf_2_txt(db_sci$myfiles, pdf2text_path="PATH/pdftotext.exe") # pdf 2 txt
db_sci$sents <- db_sci$dest \%>\% db_txt \%>\% db_sent # processing the txt transfered from pdf files and spliting into single sentences
}
