% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_txt.R
\name{db_txt}
\alias{db_txt}
\title{Processing txts transfered from pdfs, add authors and years.}
\usage{
db_txt(dest = NULL, mytxtfiles = NULL)
}
\arguments{
\item{dest}{The destination}

\item{mytxtfiles}{If there exists a txt object, you can use it directly.}
}
\description{
Processing txts transfered from pdfs, add authors and years.
}
\examples{
db_sci <- list(dest = "PATH/Mendeley_PDF/", myfiles = NULL, sents = NULL) # build a empty list comprising 3 elements. Assign path of the files saved by mendeley
db_sci$myfiles <- db_sci$dest \%>\% all_pdf # save all the pdf file names in this element
pdf_2_txt(db_sci$myfiles, pdf2text_path="PATH/pdftotext.exe") # pdf 2 txt
db_sci$sents <- db_sci$dest \%>\% db_txt \%>\% db_sent # processing the txt transfered from pdf files and spliting into single sentences
}
