% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsents.R
\name{dsents}
\alias{dsents}
\title{Extract sentences accoding a keyword or a pattern (regular expression)}
\usage{
dsents(words, sents, save_find = T, output = "output", rm_pattern = "")
}
\arguments{
\item{words}{Keywords or pattern}

\item{sents}{A sentence object.}

\item{save_find}{Logic. Save the search pattern or not.}

\item{output}{Assign a save name.}

\item{rm_pattern}{If you use a regular expresion as condition, sepcial pattern is allowed to be removed.}
}
\description{
Extract sentences accoding a keyword or a pattern (regular expression)
}
\examples{
gen_dir() # creat folders
data(demo)
#' ### single word with a variable tail
word <- "facilit"; nf <- dsents(words = paste0("\\\\b", word, "\\\\w*"), sents = db_sci$sents, output = word, rm_pattern = "")
### 后词固定+替换
word <- "estimates"; nf <- dsents(words = paste0("\\\\w*", word), sents = db_sci$sents, output = word, rm_pattern = word)
### 前词固定+替换
word <- "highly"; nf <- dsents(words = paste0(word, " \\\\w*ed"), sents = db_sci$sents, output = word,
                              rm_pattern = paste0(word, " "))
}
