% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dword_packages}
\alias{dword_packages}
\title{List all packages in the dword}
\usage{
dword_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include dword in the list?}
}
\description{
List all packages in the dword
}
\examples{
dword_packages()
}
