% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwords.R
\name{dwords}
\alias{dwords}
\title{Extract words according to Noun, Verb, Adjective and Adverb.}
\usage{
dwords(txt, savename = "result")
}
\arguments{
\item{txt}{A destination that contains txt files.}

\item{savename}{Save name}
}
\description{
Extract words according to Noun, Verb, Adjective and Adverb.
}
