% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_dir.R
\name{gen_dir}
\alias{gen_dir}
\title{Generate 5 folders name "/input", "/output", "/R", "/doc", "/fig"}
\usage{
gen_dir(...)
}
\arguments{
\item{...}{}
}
\description{
Generate 5 folders name "/input", "/output", "/R", "/doc", "/fig"
}
\examples{
gen_dir()
}
