% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_2_txt.R
\name{pdf_2_txt}
\alias{pdf_2_txt}
\title{Transefer pdf files to plain txt}
\usage{
pdf_2_txt(myfiles, pdf2text_path)
}
\arguments{
\item{myfiles}{pdf file path and name}

\item{pdf2text_path}{path of pdftotext.exe}
}
\description{
Give the pdf file path and name to this function and wait a while (depends on the file numbers), the plain txts will be yielded.
}
\examples{
filenames <- "D:/R/Mendeley_PDF/Nature Genetics/filename.pdf"
pdf2text_path <- "C:/PATH/pdftotext.exe"
pdf_2_txt(filenames, pdf2text_path)
}
