% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_freq.R
\name{word_freq}
\alias{word_freq}
\title{Get word's frequency from a txt object}
\usage{
word_freq(txt, filepath = "word_freq.csv", stop_word = NULL)
}
\arguments{
\item{txt}{A txt object}

\item{filepath}{Assign a output path. The default is "word_freq.csv" under root folder.}

\item{stop_word}{Character vector.}
}
\description{
Get word's frequency from a txt object
}
