% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_sent.R
\name{db_sent}
\alias{db_sent}
\title{Sliping sentences from txt object.}
\usage{
db_sent(txt = txt, minL = 50, maxL = 800)
}
\arguments{
\item{txt}{A txt object containg characters.}

\item{minL}{Filter the sentence by length longer than minL.}

\item{maxL}{Filter the sentence by length shorter than maxL.}
}
\description{
Sliping sentences from txt object.
}
