% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_txt.R
\name{db_txt}
\alias{db_txt}
\title{Processing txts transfered from pdfs, add authors and years.}
\usage{
db_txt(dest = NULL, mytxtfiles = NULL)
}
\arguments{
\item{dest}{The destination}

\item{mytxtfiles}{If there exists a txt object, you can use it directly.}
}
\description{
Processing txts transfered from pdfs, add authors and years.
}
