% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsents.R
\name{dsents}
\alias{dsents}
\title{Extract sentences accoding a keyword or a pattern (regular expression)}
\usage{
dsents(words, sents, save_find = T, output = "output", rm_pattern = "")
}
\arguments{
\item{words}{Keywords or pattern}

\item{sents}{A sentence object.}

\item{save_find}{Logic. Save the search pattern or not.}

\item{output}{Assign a save name.}

\item{rm_pattern}{If you use a regular expresion as condition, sepcial pattern is allowed to be removed.}
}
\description{
Extract sentences accoding a keyword or a pattern (regular expression)
}
