% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwords.R
\name{dwords}
\alias{dwords}
\title{Extract words according to Noun, Verb, Adjective and Adverb.}
\usage{
dwords(txt, path_TreeTagger, savename = "result")
}
\arguments{
\item{txt}{A destination that contains txt files.}

\item{path_TreeTagger}{Path of TreeTagger, TreeTagger is used to lemma words.}

\item{savename}{Save name}
}
\description{
Extract words according to Noun, Verb, Adjective and Adverb.
}
\examples{
gen_dir() # creat folders
data(demo)
txts <- dwords(db_sci$sents, path_TreeTagger="PATH/TreeTagger", "outname")
}
